remove test_plot.py and test_operationHeatMap in favor of new test_plotFunctions.py
